Make zfslocalpv client testable
